test-requirements.txt
charms.reactive
flake8
stestr
git+https://github.com/juju/charm-helpers
git+https://github.com/openstack/charms.openstack.git#egg=charms.openstack

